Move ir-related code into myia.ir submodule
