+++ myia/debug/buche.py
@@ -1,5 +1,5 @@
 """Provide a buche function that works on Myia objects."""
 
 # Load custom hrepr methods for Graph etc.
-from myia.debug import gprint  # noqa
+from . import gprint  # noqa
 from buche import buche  # noqa
