+++ myia/debug/do_inject.py
@@ -1,7 +1,7 @@
 """Inject functionality in all modules, only import for debugging purposes."""
 
 
-from myia.debug.inject import inject_suite
+from .inject import inject_suite
 
 
 inject_suite()
