+++ myia/debug/label.py
@@ -1,11 +1,11 @@
 """Utilities to generate or map labels for nodes and graphs."""
 
 
-from myia.info import DebugInfo
-from myia.anf_ir import Graph
-from myia.anf_ir_utils import \
+from ..info import DebugInfo
+from ..ir.anf import Graph
+from ..ir.utils import \
     is_constant, is_constant_graph, is_parameter, is_special
-from myia.prim import Primitive
+from ..prim import Primitive
 
 
 short_relation_symbols = {
