+++ myia/ir/__init__.py
@@ -1,0 +1,14 @@
+"""Main exports for ir submodule."""
+
+from .abstract import Node  # noqa
+from .anf import Graph, ANFNode, Apply, Constant, Parameter, Special  # noqa
+from .clone import GraphCloner  # noqa
+from .utils import (  # noqa
+    succ_deep, succ_deeper, succ_incoming, succ_bidirectional,
+    exclude_from_set, freevars_boundary,
+    dfs, toposort, accessible_graphs,
+    destroy_disconnected_nodes,
+    isomorphic,
+    replace,
+    is_apply, is_parameter, is_constant, is_constant_graph, is_special
+)
