+++ myia/ir/abstract.py
