+++ myia/ir/utils.py
@@ -1,9 +1,11 @@
 """Utilities for manipulating and inspecting the IR."""
+
 from typing import Any, Iterable, Callable, Set
 
-from myia.anf_ir import ANFNode, Apply, Constant, Graph, Parameter, Special
-from myia.graph_utils import dfs as _dfs, toposort as _toposort, \
+from ..graph_utils import dfs as _dfs, toposort as _toposort, \
     FOLLOW, NOFOLLOW, EXCLUDE
+
+from .anf import ANFNode, Apply, Constant, Graph, Parameter, Special
 
 
 #######################
