+++ myia/unify.py
@@ -1,8 +1,9 @@
 """Unification module."""
+
 from typing import Dict, List, Iterable, Union, Callable, TypeVar, Type, Any
 from functools import reduce
 
-from myia.utils import Registry
+from .utils import Registry
 
 T = TypeVar('T')
